refactor: Consolidate duplicated model picker item logic
